requirements.txt
pyyaml==6.0
orjson==3.8.3